+++ src/OpenTelemetry.Instrumentation.GrpcNetClient/Implementation/GrpcInstrumentationEventSource.cs
@@ -49,4 +49,25 @@ internal sealed class GrpcInstrumentationEventSource : EventSource
     {
         this.WriteEvent(3, handlerName, eventName, ex);
     }
+
+    [Event(4, Message = "Request is filtered out. HandlerName: '{0}', EventName: '{1}', OperationName: '{2}'.", Level = EventLevel.Verbose)]
+    public void RequestIsFilteredOut(string handlerName, string eventName, string operationName)
+    {
+        this.WriteEvent(4, handlerName, eventName, operationName);
+    }
+
+    [Event(5, Message = "Filter threw exception, request will not be collected. HandlerName: '{0}', EventName: '{1}', OperationName: '{2}', Exception: {3}.", Level = EventLevel.Error)]
+    public void RequestFilterException(string handlerName, string eventName, string operationName, string exception)
+    {
+        this.WriteEvent(5, handlerName, eventName, operationName, exception);
+    }
+
+    [NonEvent]
+    public void RequestFilterException(string handlerName, string eventName, string operationName, Exception ex)
+    {
+        if (this.IsEnabled(EventLevel.Error, EventKeywords.All))
+        {
+            this.RequestFilterException(handlerName, eventName, operationName, ex.ToInvariantString());
+        }
+    }
 }